NetCDF: Use `pkg-config` to find libs
The current logic only adds `libnetcdf` to the link line, but no
transitive libs.

Both `nc-config` and `pkg-config` provide this information to
populate transitive libs and thus symbols during linking. We pick
the latter, since it will also work when you cross-compile for
a different architecture.

Here is the helper function in CMake that we use to query
`pkg-config --libs` et al.:
  https://cmake.org/cmake/help/latest/module/FindPkgConfig.html