Overhaul InputSoundingData read and calc_rho_p routines
Now, all *_inp_sound data (and correspondingly, the *_integ data) will
range from k=0 at the surface to the domain top. This provides thermo
quantities at the boundaries and also simplifies the hse integration in
calc_rho_p().

Note that unless the input sounding exactly includes a point
corresponding to the top of the simulation domian, the value at ztop
will either be interpolated or extrapolated.