Fix indexing, parallel behavior
Note: The Rayleigh damping implementation will differ slightly from
before in the case without terrain. Previously, ztop was set to the top
of the domain; now, ztop is set to max(z_phys_cc[:,:,khi]) -- where ztop
is the value at which the sine-squared damping factor is one. Therefore,
the top layer of cells will now have a factor of exactly one instead of
nearly one.