Set MOST ghost cell by extrapolating from first cell
This is how the surface BC is set in SOWFA and is a simplification of Moeng
1984, in which the first-cell velocity gradient -- used to update the tke
production term -- is determined by interpolating between the resolved velocity
gradient at the first zface and the modeled velocity gradient at a test height
below the first cell center. We neglect the latter contribution for now.